Move warmups into shared
